Restructure helpers folder. Add csv parser
