+++ warehouse/products/helpers/__init__.py
@@ -1,0 +1,12 @@
+from warehouse.products.helpers.utils import (
+    ProductDTO,
+    ProductPriceDTO,
+    ProductDTOService,
+    ProductPriceDTOService,
+    DtoManager,
+    remove_stock,
+    add_stock,
+    possibleparts_select_list,
+    suppliers_select_list,
+)
+from warehouse.products.helpers.importers import ProductPair, StockImporter, StockParser
